I want to add qrcode scanner

+++ app/Http/Controllers/staffs/attendanceController.php
@@ -61,17 +61,16 @@ class attendanceController extends Controller
 
 
         $qrcode = str_replace(" ", "Faculty::", $course->faculty->faculty."_".$course->department->dept."_". $course->title."_".$course->code).".png";
-        $qrcodes = str_replace(" ", "-", $course->faculty->faculty."_".$course->department->dept."_". $course->title."_".$course->code)."|";
         $attendance->qrcode = $qrcode;
 
         $attendance->save();
         // return $attendance->id;
-        $message = "Faculty:".$course->faculty->faculty."<br>";
-        $message .= "Department:".$course->department->dept."<br>";
-        $message .= "Course title:".$course->title."<br>";
-        $message .= "Course code:".$course->code."<br>";
-        $message .= "Course unit:".$course->unit."<br>";
-        $message .= "Attendance date:".$request->date."<br>";
+        $message = "Faculty:".$course->faculty->faculty."\n";
+        $message .= "Department:".$course->department->dept."\n";
+        $message .= "Course title:".$course->title."\n";
+        $message .= "Course code:".$course->code."\n";
+        $message .= "Course unit:".$course->unit."\n";
+        $message .= "Attendance date:".$request->date."\n";
         $message .= "Smartcode:smartme".time()."_".$attendance->id;
         QrCode::size(500)->format('png')->generate(html_entity_decode($message), public_path('qrcode/'.$qrcode));
         return redirect()->back()->with('success', 'New attendance created successfully');
@@ -118,12 +117,24 @@ class attendanceController extends Controller
         $attendance->faculty_id = $request->faculty;
         $attendance->date = $request->date;
         $attendance->level = $request->level;
-        unlink(public_path("qrcode/".$attendance->qrcode));
+
+        if($attendance->qrcode){
+            unlink(public_path("qrcode/".$attendance->qrcode));
+        }
         $course = Course::with(['faculty', 'department'])->where('id', $request->course)->firstOrFail();
-        $qrcode = str_replace(" ", "-", $course->faculty->faculty."_".$course->department->dept."_". $course->title."_".$course->code). time() .".png";
+       $qrcode = str_replace(" ", "-", $course->faculty->faculty."_".$course->department->dept."_". $course->title."_".$course->code). time() .".png";
         $qrcodes = str_replace(" ", "-", $course->faculty->faculty."_".$course->department->dept."_". $course->title."_".$course->code)."|";
         $attendance->qrcode = $qrcode;
         $attendance->update();
+        // return $attendance->id;
+        $message = "Faculty:".$course->faculty->faculty."\n";
+        $message .= "Department:".$course->department->dept."\n";
+        $message .= "Course title:".$course->title."\n";
+        $message .= "Course code:".$course->code."\n";
+        $message .= "Course unit:".$course->unit."\n";
+        $message .= "Attendance date:".$request->date."\n";
+        $message .= "Smartcode:smartme".time()."_".$attendance->id;
+        QrCode::size(500)->format('png')->generate(html_entity_decode($message), public_path('qrcode/'.$qrcode));
         // return $attendance->id;
         // QrCode::size(500)->format('png')->generate($qrcodes."smartqrcode".$attendance->id, public_path('qrcode/'.$qrcode));
         return redirect()->back()->with('success', $course->title." || ".$course->code.' attendance updated successfully');
@@ -140,7 +151,9 @@ class attendanceController extends Controller
     {
         //
         $attendance = Attendance::with(['course'])->where('id', $id)->firstOrFail();
-        unlink(public_path("qrcode/".$attendance->qrcode));
+        if($attendance->qrcode){
+            unlink(public_path("qrcode/".$attendance->qrcode));
+        }
         $attendance->forceDelete();
 
         return redirect()->back()->with('success', 'Attendance '. $attendance->course->title.' deleted successfully');
